organized data type for interfaces and contants in the same folder/category
