+++ src/components/Editor/Editor.FunctionalComponent
@@ -7,7 +7,7 @@ import {
   Form
 } from "react-bootstrap";
 
-import Canvas, { CanvasController, CanvasOrderDirection } from "../Canvas/Canvas";
+import Canvas, { CanvasOrderDirection } from "../Canvas/Canvas";
 import "./Editor.css";
 
 import { fabric } from "fabric";
@@ -22,9 +22,10 @@ import {
   DEFAULT_FONT,
   DEFAULT_TSHIRT_ID,
   DEFAULT_FILL_COLOR
-} from "../../config/constants";
+} from "../../data_type/constants";
 
-import State from "../../interfaces/State";
+import State from "../../data_type/interfaces/State";
+import CanvasController from "../../data_type/interfaces/CanvasController";
 
 interface Props {
   editor: State;
