+++ src/components/ObjectControlButtonsGroup/ObjectControlButtonsGroup.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { CanvasOrderDirection } from "../Canvas/Canvas";
 import {
   Button,
   ButtonGroup,
 } from "react-bootstrap";
-import State from "../../interfaces/State";
+import {State} from "../../data_type/interfaces";
+import { CanvasOrderDirection } from '../../data_type/constants';
 
 interface Props {
   // You can define props here if needed
